# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestión E-commerce",
        "description": "API para gestión de órdenes de venta de un e-commerce",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products/{id}/restore": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Restaurar un producto eliminado lógicamente",
                "description": "Restaura un producto que fue eliminado mediante soft delete. El producto vuelve a estar disponible para consultas normales. Requiere autenticación. Solo funciona con productos que han sido eliminados lógicamente. Invalida automáticamente el cache.",
                "operationId": "cac6708948b75b3fcc612aed6626824c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID único del producto eliminado que se desea restaurar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto restaurado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto restaurado exitosamente"
                                        },
                                        "product": {
                                            "$ref": "#/components/schemas/ProductDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado o no está eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Product] 1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/{id}/force": {
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Eliminar permanentemente un producto",
                "description": "Elimina un producto de forma permanente de la base de datos. Esta acción es irreversible. Requiere autenticación. Solo funciona con productos que ya han sido eliminados lógicamente (soft deleted). Invalida automáticamente el cache.",
                "operationId": "3c95bd58069e5b3d7356552bc209e53a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID único del producto eliminado lógicamente que se desea eliminar permanentemente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado permanentemente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto eliminado permanentemente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado o no está eliminado lógicamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Product] 1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ProductDetail": {
                "title": "Product Detail",
                "description": "Modelo detallado de producto del e-commerce",
                "required": [
                    "id",
                    "name",
                    "description",
                    "price",
                    "stock"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del producto",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre del producto",
                        "type": "string",
                        "maxLength": 255,
                        "example": "iPhone 15 Pro"
                    },
                    "description": {
                        "description": "Descripción detallada del producto",
                        "type": "string",
                        "example": "Último modelo de iPhone con chip A17 Pro"
                    },
                    "price": {
                        "description": "Precio del producto en USD",
                        "type": "number",
                        "format": "float",
                        "minimum": 0.01,
                        "example": 1199.99
                    },
                    "stock": {
                        "description": "Cantidad disponible en inventario",
                        "type": "integer",
                        "minimum": 0,
                        "example": 50
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Fecha y hora de última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:30:45.000000Z"
                    },
                    "deleted_at": {
                        "description": "Fecha y hora de eliminación lógica (null si está activo)",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}